Extract ValidatedNumberInput
